Created src/assets and moved the images folder into src/assets and also  moved all the styling for individual components to their respective folders
