+++ public/css/style.css
@@ -1,11 +1,3 @@
-* {
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  font-family: 'Roboto', sans-serif;
-}
-
 .container-mask {
   display: -webkit-box;
   display: -ms-flexbox;
@@ -63,4 +55,12 @@
   width: 80px;
   font-size: 0.6rem;
   font-weight: bold;
+}
+
+* {
+  -webkit-box-sizing: border-box;
+          box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-family: 'Roboto', sans-serif;
 }
