+++ src/assets/images/plateau-logo.png
